LPD-71240 Customizing rootModelHierarchy with the new pattern
