+++ modules/apps/object/object-rest-impl/src/main/java/com/liferay/object/rest/internal/jaxrs/feature/ObjectFeature.java
@@ -5,13 +5,11 @@
 
 package com.liferay.object.rest.internal.jaxrs.feature;
 
-import com.liferay.object.rest.internal.jaxrs.container.request.filter.NestedFieldsContainerRequestFilter;
 import com.liferay.object.rest.internal.jaxrs.param.converter.provider.ScopeKeyParamConverterProvider;
 import com.liferay.object.service.ObjectDefinitionLocalService;
 import com.liferay.object.service.ObjectRelationshipLocalService;
 import com.liferay.portal.kernel.service.GroupLocalService;
 
-import jakarta.ws.rs.Priorities;
 import jakarta.ws.rs.core.Feature;
 import jakarta.ws.rs.core.FeatureContext;
 
@@ -33,10 +31,6 @@ public class ObjectFeature implements Feature {
 
 	@Override
 	public boolean configure(FeatureContext featureContext) {
-		featureContext.register(
-			new NestedFieldsContainerRequestFilter(
-				_objectDefinitionLocalService, _objectRelationshipLocalService),
-			Priorities.USER + 10);
 		featureContext.register(
 			new ScopeKeyParamConverterProvider(_groupLocalService));
